backend/requirements.txt
flask==2.3.3
python-docx==0.8.11
gunicorn==20.0.4
